1-6章课上代码/课上代码/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.2
distro==1.9.0
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.0.0
langchain-openai==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langsmith==0.4.10
MarkupSafe==2.1.5
modelscope==1.29.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
openai==1.98.0
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pillow==11.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
setuptools==70.2.0
sniffio==1.3.1
sympy==1.13.3
tenacity==9.1.2
tiktoken==0.9.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
xxhash==3.5.0
zstandard==0.23.0
